src/app/_services/adapters/http-adapter/http-adapter.service.ts
import { UserCredentials } from './../../../_types/user-credentials.type';
import { HttpClient } from '@angular/common/http';
import { Injectable } from '@angular/core';
import { Observable } from 'rxjs';

@Injectable({
  providedIn: 'root'
})
export class HttpAdapterService {

  constructor(
    private httpClient: HttpClient,
  ) { }

  patch<T>(url: string, body: any): Promise<T> {
    return this.httpClient.patch<T>(url, body).toPromise();
  }

  post<T>(url: string, body: any): Promise<T> {
    return this.httpClient.post<T>(url, body).toPromise();
  }

  get<T>(url: string): Observable<T> {
    return this.httpClient.get<T>(url);
  }
}
